+++ includes/structure/header.php
@@ -70,6 +70,7 @@ function bfg_security_headers() {
 
 	send_frame_options_header();
 	header( 'X-Content-Type-Options: nosniff' );
+	header( 'X-XSS-Protection: 1; mode=block' );
 
 }