+++ js/scripts.js
@@ -1,3 +1,6 @@
+import im from '../node_modules/include-media-export/include-media.js';
+import fitvids from '../node_modules/fitvids/dist/fitvids.js';
+
 (function() {
 	'use strict';